% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBNApaper.patients.r
\docType{data}
\name{CBNApaper.patients}
\alias{CBNApaper.patients}
\title{TCGA-BRCA barcodes of the patients analised in CBNA(2019)}
\format{
A \code{character} vector with 747 barcodes corresponding to the TCGA-BRCA
patients analysed in CBNA study.
}
\usage{
CBNApaper.patients
}
\description{
A \code{character} vector containing the barcode used for the
analysis described in
\href{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1007538}{CBNA}
study. This study was selected for comparison purposes to assess the performance of
our \emph{DynamicCancerDriver} method and several popular methods for cancer driver
inference.
}
\references{
CBNA: A control theory based method for identifying coding and
non-coding cancer drivers\cr
Pham VVH, Liu L, Bracken CP, Goodall GJ, Long Q, et al. (2019) \cr
PLOS Computational Biology 15(12): e1007538. \cr
https://doi.org/10.1371/journal.pcbi.1007538
}
\keyword{CBNA}
\keyword{Cancer_Driver}
\keyword{TCGA-BRCA}
