% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGC.driverNames.r
\docType{data}
\name{CGC.driverNames}
\alias{CGC.driverNames}
\title{Cancer Gene Consensus (v94, n=719)}
\format{
A \code{dataframe} with 719 rows and 5 columns.
}
\usage{
CGC.driverNames
}
\description{
\code{Dataframe} containing the names of the 719 cancer driver genes in
the Cancer Gene Census (v94). The \code{dataframe} has 5 variables:
\itemize{
 \item{\code{IDs: }}{ The original name of the driver obtained
 from \url{https://cancer.sanger.ac.uk/census}}
 \item{\code{Ensembl.ID: }}{ Ensembl ID of the cancer driver}
 \item{\code{HGNC.ID: }}{ HGNC ID of the cancer driver}
 \item{\code{HGNC.symbol: }}{ (updated) Hugo symbol of the cancer driver}
 \item{\code{NCBI.ID: }}{ NCBI ID of the cancer driver}
}
}
\references{
\url{https://cancer.sanger.ac.uk/cosmic}
}
\keyword{CGC}
\keyword{Cancer_Driver}
