% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePPIRank.r
\name{CreatePPIRank}
\alias{CreatePPIRank}
\title{CreatePPIRank}
\usage{
function(geneIDs = NULL, PPImatrix = NULL)\cr
}
\arguments{
\item{geneIDs}{A \code{character} vector containing genes IDs.IDs can be
Ensembl.ID (recommended), HGNC.ID, NCBI.ID or HGNC.symbol}

\item{PPImatrix}{A 2 column \code{matrix} containing the input nodes (column 1)
, and the output nodes (column 2) oof a PPI network. If \code{NULL} (default),
the PPI network decribed in
\url{https://www.science.org/doi/10.1126/scisignal.2001699} is used as reference.}
}
\value{
A \code{dataframe} following variables:
  \enumerate{
          \item{\code{IDs}}{Ensembl.ID, HGNC.ID, NCBI.ID or HGNC.symbol of the genes
          in the PPI network}
          \item{\code{n.in}}{Count of the gene as input node}
          \item{\code{n.out}}{Count of the gene as output node}
          \item{\code{total}}{Count of the total times the gene appears in the PPI network}
          }
}
\description{
CreatePPIRank takes each gene in \code{geneIDs} and determines
the number of times such a gene appears as input node (\code{n.in})
and output node (\code{n.out}) in the PPI network described by \code{PPImatrix}.
If no PPImatrix is provided, the PPI network described at
\url{https://www.science.org/doi/10.1126/scisignal.2001699} is used as reference.
The rank corresponds to the \code{total} count of the gene in the PPI network.
}
\examples{
\dontrun{
   data("GSE75688_TPM_tumor", package = "DynamicCancerDriver")
   CreatePPIRank(colnames(GSE75688_TPM_tumor)[1:100])
}

}
\references{

}
\seealso{
\link[DynamicCancerDriver]{findDCD}
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
