% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCovariate.r
\name{findCovariate}
\alias{findCovariate}
\title{findCovariate}
\usage{
function(GeneExpression, FS, PPIrank=NULL)
}
\arguments{
\item{GeneExpression}{A \code{dataframe} (preferred) or a \code{matrix} containing mRNA gene expression.\cr
Columns represent mRNAs and rows represent samples. Column names should correspond
to gene IDs. Gene IDs can be from any of the following nomenclatures:
\emph{Ensembl.ID} (recommended), \emph{HGNC.ID, NCBI.ID} or \emph{HGNC.symbol}.}

\item{FS}{A \code{character} vector containing the names of the features
for which a covariate is to be found.}

\item{PPIrank}{(optional) A \code{dataframe} obtained from
\code{\link{CreatePPIRank}}}
}
\value{
A \code{dataframe} with the following two variables:
  \enumerate{
          \item{\emph{Feature: }}{The vector \code{FS} of features.
          The name of this variable will be \emph{Ensembl.ID, HGNC.ID, NCBI.ID}
          or \emph{HGNC.symbol}}.
          \item{\code{scontrol: }}{For each \emph{Feature}, the name of the non-PPI gene with the
          largest Pearson correlation}
          }
}
\description{
For each feature in \code{FS}, \code{findCovariate}
 finds the non-PPI gene (in \code{GeneExpression}) with the largest Pearson correlation with the feature.\cr
 For a proper functioning, the data of all features in \code{FS} must be included as columns in
 \code{GeneExpression}.
}
\examples{
\dontrun{
   data("GSE75688_TPM_tumor", package = "DynamicCancerDriver")
   FS <- colnames(GSE75688_TPM_tumor)[1:100]

   sControl <- findCovariate(GeneExpression = GSE75688_TPM_tumor[,1:500]
   , FS = FS)
   }


}
\references{

}
\seealso{
\link[DynamicCancerDriver]{findDCD}
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
