% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findZ.r
\name{findZ}
\alias{findZ}
\title{findZ}
\usage{
function(GeneExpression, FS, pathCovariate, elbo_tol = 1e-3)\cr
}
\arguments{
\item{GeneExpression}{A \code{matrix} containing mRNA gene expression.\cr
Columns represent mRNAs, rows represent samples. Column names can be any of
the following: Hugo Gene Symbol (HGNC.symbol), Hugo Gene ID (HGNC.ID),
EnsemblID, NBCI.}

\item{FS}{A \code{character} vector containing the names of the features
to be used for the calculation of the pseudotime score.}

\item{pathCovariate}{A \code{named vector} containing the data of a path
covariate.}

\item{elbo_tol}{A \code{numeric} value. The relative pct change in the
evidence lower bound (ELBO) below which \code{phenopath} calculation is
considered converged.}
}
\value{
A \code{dataframe} with the following two variables:
  \enumerate{
          \item{\code{Feature:}}{The vector \code{FS} of features}
          \item{\code{scontrol:}}{The name of the non-PPI gene with the
          largest Pearson correlation with the respective feature}
          }
}
\description{
\code{phenopath} package
}
\examples{
\dontrun{
   data("GSE75688_TPM_tumor", package = "DynamicCancerDriver")
   FS <- colnames(GSE75688_TPM_tumor)[1:100]
   GE <- GSE75688_TPM_tumor[,1:500]
   sControl <- findCovariate(GeneExpression = GSE75688_TPM_tumor[,1:500]
   , FS = FS)

   #toy example, using "VIM" as path covariate
   z <- findZ(GeneExpression = GE
             , FS, pathCovariate = GSE75688_TPM_tumor[,"ENSG00000026025"]
             , elbo_tol = 1e-3)
             }


}
\references{

}
\seealso{
\link[DynamicCancerDriver]{findDCD}
, \link[DynamicCancerDriver]{parallelCI}
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
