% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE75688_sample_information.r
\docType{data}
\name{GSE75688_sample_information}
\alias{GSE75688_sample_information}
\title{Information of the samples in GSE75688_TPM dataset}
\format{
A \code{dataframe} with 528 observations (rows) and 5 variables (columns).
}
\usage{
GSE75688_sample_information
}
\description{
A \code{dataframe} containing the sample ID ("sample"),
type of sample ("type") that can be single cell ("SC") of bulk data ("Bulk"), and
the kind of sample cells ("index", "index2", "index3").\cr
"index" can be either, "Tumor",  or  "nonTumor".\cr
"index2" can be "Tumor", "Stromal", or "Immune".\cr
"index3" can be "Tumor", "Stromal", "Myeloid", "Tcell", "Bcell", or "Immune".
}
\references{
Chung, W., Eum, H., Lee, HO. et al. Single-cell RNA-seq enables
comprehensive tumour and immune cell profiling in primary breast cancer. \cr
Nat Commun 8, 15081 (2017). https://doi.org/10.1038/ncomms15081
}
\keyword{Single_Cell}
\keyword{dataset}
