% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE75688_TPM.r
\docType{data}
\name{GSE75688_TPM}
\alias{GSE75688_TPM}
\title{Gene expression data (TPM) from GSE75688}
\format{
A \code{matrix} with 563  rows and 57915 columns.
samples are represented in rows while features (genes) in columns.
}
\usage{
GSE75688_TPM
}
\description{
Single cell RNA sequencing (RNA-seq) for 549 primary breast cancer
cells and lymph node metastases from 11 patients with distinct molecular subtypes
(BC01-BC02, estrogen receptor positive (ER+); BC03, double positive (ER+ and HER2+)
; BC03LN, lymph node metastasis of BC03; BC04-BC06, human epidermal growth factor
receptor 2 positive (HER2+); BC07-BC11, triple-negative breast cancer (TNBC);
BC07LN, lymph node metastasis of BC07) and matched bulk tumors.
}
\references{
Chung, W., Eum, H., Lee, HO. et al. Single-cell RNA-seq enables
comprehensive tumour and immune cell profiling in primary breast cancer. \cr
Nat Commun 8, 15081 (2017). https://doi.org/10.1038/ncomms15081
}
\keyword{Single_Cell}
\keyword{dataset}
