% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE75688_TPM_tumor.r
\docType{data}
\name{GSE75688_TPM_tumor}
\alias{GSE75688_TPM_tumor}
\title{Gene expression data (TPM) from GSE75688 (only SC tumor samples)}
\format{
A \code{matrix} with 317 rows (samples) and 9551 columns (genes).
}
\usage{
GSE75688_TPM_tumor
}
\description{
A \code{matrix} containing the single cell RNA sequencing (RNA-seq)
from GSE75688_TPM after filtering and pre-processing. A pre process was
performed to samples from tumor cells. A filtering process was perfomed to discard
the gene expression of genes not expressed in at east 20% of the samples.\cr
This dataset is used for the experiment described in our paper.
}
\references{
Chung, W., Eum, H., Lee, HO. et al. Single-cell RNA-seq enables
comprehensive tumour and immune cell profiling in primary breast cancer. \cr
Nat Commun 8, 15081 (2017). https://doi.org/10.1038/ncomms15081
}
\keyword{Single_Cell}
\keyword{dataset}
