% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelCI.r
\name{parallelCI}
\alias{parallelCI}
\title{parallelCI}
\usage{
function(GeneExpression,sControl,z,eventAt
, chunk_size = 50, returnModel = F)\cr
}
\arguments{
\item{GeneExpression}{A \code{matrix} containing mRNA gene expression.\cr
Columns represent mRNAs, rows represent samples. Column names can be any of
the following: Hugo Gene Symbol (HGNC.symbol), Hugo Gene ID (HGNC.ID),
EnsemblID, NBCI.}

\item{sControl}{A 2 column matrix containing a gene ID (1st column) and 1 non-PPI
gene to be uses as covariate for the \code{CausalImpact} model. For a correct
functioning, the pseudotime ordered data of all element in \code{sControl} need
to be included as columns in \code{GeneExpression} matrix.}

\item{z}{A numeric vector containing the pseudotime score used for ordering
the samples in \code{GeneExpression}. For a correct functioning, \code{z} needs to
follow ascending order and this order must agree with the order of the samples (rows)
of the \code{GeneExpression} matrix.}

\item{eventAt}{An integer with the index (in pseudotime order) of the sample where
the most significant change is inferred to happen.}

\item{chunk_size}{An integer indicating the number of genes to be passed to
each worker during the parallel calculation. (50 by default)}

\item{returnModel}{A boolean. If \code{TRUE}, the full causal impact model
(as calculated by \code{CausalImpact} package) is returned. if \code{FALSE}
(default), only the main parameters of the \code{CausalImpact} are returned,}
}
\value{
A \code{list} where each element is the full \code{CausalImpact} model
(if \code{returnModel = TRUE}) or the simplified \code{CausalImpact} model
(if \code{returnModel = FALSE}) of one gene in the 1st column of \code{sControl}.
}
\description{
parallelCI uses \code{parallel} package for implementing
a parallelised calculation of the \code{CausalImpact} on gene expression.
It is assumed that the \code{GeneExpression} matrix provided contains
pseudotime ordered gene expression. \code{z} is the pseudotime score used for
ordering the gene expression, and \code{eventAt} indicates the sample at
the most significant change occurs.
}
\examples{
\dontrun{
   data("GSE75688_TPM_tumor", package = "DynamicCancerDriver")
   FS <- colnames(GSE75688_TPM_tumor)[1:100]
   GE <- GSE75688_TPM_tumor[,1:500]
   sControl <- findCovariate(GeneExpression = GSE75688_TPM_tumor[,1:500]
   , FS = FS)

   #toy example, using "VIM" as path covariate
   z <- findZ(GeneExpression = GE
             , FS, pathCovariate = GSE75688_TPM_tumor[,"ENSG00000026025"]
             , elbo_tol = 1e-3)
   GE <- GE[order(z),,drop=F]
   z <- z[order(z),1, drop=F]

   parCI <-parallelCI(GE,sControl,z,eventAt=7)

}

}
\references{

}
\seealso{
\link[DynamicCancerDriver]{findCovariate},
\link[DynamicCancerDriver]{parallelCI}
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
