% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI.r
\docType{data}
\name{PPI}
\alias{PPI}
\title{Protein-protein interaction network}
\format{
A \code{dataframe} with 34814 observations (rows) and 5 variables (columns).
samples are represented in rows while features (genes) in columns. The variables are
"Input-node Gene Symbol", "Input-node GeneID", "Output-node Gene Symbol",
"Output-node GeneID", and "Edge direction score" respectively.
}
\usage{
PPI
}
\description{
A \code{dataframe} containing the information of the PPI network
described by
\href{https://www.science.org/doi/10.1126/scisignal.2001699}{Vinayagam et al. (2011)}
}
\references{
Vinayagam, A., Stelzl, U., Foulle, R., Plassmann, S., Zenkner, M., Timm, J., Assmus,
H. E., Andrade-Navarro, M. A., andWanker, E. E. \cr
(2011). A directed protein interaction network for investigating intracellular signal transduction.\cr
Science signaling, 4(189).
}
\keyword{PPI}
