% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_BRCA_TP_NormCounts.r
\docType{data}
\name{TCGA_BRCA_TP_NormCounts}
\alias{TCGA_BRCA_TP_NormCounts}
\title{TCGA_BRCA Normalised Gene Expression Counts.}
\format{
A \code{matrix} with 1101 samples (rows) and 23192 genes (columns).
}
\usage{
TCGA_BRCA_TP_NormCounts
}
\description{
A \code{matrix} containing the gene expression (normalised counts)
of the TCGA-BRCA project (dataset downloaded in Aug, 2021). The dataset download and
the normalisation process were perfomed by using the
\href{https://bioconductor.org/packages/release/bioc/html/TCGAbiolinks.html}{TCGABiolinks} package. \cr
This dataset is used for the bechmarking analysis described in our paper.
}
\references{

}
\keyword{BRCA}
\keyword{TCGA}
\keyword{dataset}
