% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBNApaper.patients.r, R/CGC.driverNames.r
\docType{data}
\name{CBNApaper.patients}
\alias{CBNApaper.patients}
\title{TCGA-BRCA barcodes of the patients analised in CBNA(2019)}
\format{
A \code{character} vector with 747 barcodes corresponding to the TCGA-BRCA
patients analysed in CBNA study.

A \code{dataframe} with 719 rows and 5 columns.
}
\usage{
CBNApaper.patients

CGC.driverNames
}
\description{
\code{character} vector containing the barcodes used for analisys in
\href{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1007538}{CBNA}
study. This study was selected for comparison purposes to assess the performance of
our \emph{DynamicCancerDriver} method and several popular methods for cancer driver
inference.


}
\references{
CBNA: A control theory based method for identifying coding and
non-coding cancer drivers\cr
Pham VVH, Liu L, Bracken CP, Goodall GJ, Long Q, et al. (2019) \cr
PLOS Computational Biology 15(12): e1007538. \cr
https://doi.org/10.1371/journal.pcbi.1007538

\url{https://cancer.sanger.ac.uk/cosmic}
}
\keyword{CBNA}
\keyword{CGC}
\keyword{Cancer_Driver}
\keyword{TCGA-BRCA}
