% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCovariate.r
\name{findCovariate}
\alias{findCovariate}
\title{findCovariate}
\usage{
function(GeneExpression, FS, PPIrank=NULL)\cr
}
\arguments{
\item{GeneExpression}{A \code{matrix} containing mRNA gene expression.\cr
Columns represent mRNAs, rows represent samples. Column names can be any of
the following: Hugo Gene Symbol (HGNC.symbol), Hugo Gene ID (HGNC.ID),
EnsemblID, NBCI.}

\item{FS}{A \code{character} vector containing the names of the features
for which a covariate is to be found.}

\item{PPIrank}{(optional) A \code{dataframe} obtained from \code{\link{CreatePPIRank}}}
}
\value{
A \code{dataframe} with the following two variables:
  \enumerate{
          \item{\code{Feature:}}{The vector \code{FS} of features}
          \item{\code{scontrol:}}{The name of the non-PPI gene with the
          largest Pearson correlation with the respective feature}
          }
}
\description{
For each feature in \code{FS} \code{findCovariate}
finds the non-PPI gene with the largest pearson correlation with the feature.
The data of all features in \code{FS}, need to be included in the
\code{GeneExpression} parameter.
}
\examples{
\dontrun{
   data("GSE75688_TPM_tumor", package = "DynamicCancerDriver")

----- Find Dynamic Cancer Drivers, PPI top 40\% -----
DCD.HER2time_SC <- findDCD(GeneExpression = GSE75688_TPM_tumor
                           , pathCovariate = "HER2"
                           , PPItop = 0.3
                           , findEvent = TRUE)
}

}
\references{

}
\seealso{
\link[DynamicCancerDriver]{findCovariate},
\link[DynamicCancerDriver]{parallelCI}
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
