% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDCD.r
\name{findDCD}
\alias{findDCD}
\title{findDCD}
\usage{
function(GeneExpression, z=NULL, pathCovariate =NULL
, findEvent = T, Step=1, chunk_size= 100
, PPItop = 0.3, alpha=0.05, CIniter=200
, returnModel=F, elbo_tol=1e-3, project = NULL)
}
\arguments{
\item{GeneExpression}{A \code{dataframe} (preferred) or a \code{matrix} containing mRNA gene expression.\cr
Columns represent mRNAs and rows represent samples. Column names should correspond
to gene IDs. Gene IDs can be from any of the following nomenclatures:
\emph{Ensembl.ID} (recommended), \emph{HGNC.ID, NCBI.ID} or \emph{HGNC.symbol}.}

\item{z}{A \code{numeric vector} with pseudotime score for each sample.
If \code{NULL} (default), pseudotime score is calculated
by using \href{https://www.bioconductor.org/packages/release/bioc/html/phenopath.html}{phenopath} package
and the \code{pathCovariate}.}

\item{findEvent}{If \code{TRUE} (default) samples are ordered in pseudotime order
and deviations from normal to cancerogenic are assessed by using the
\code{CausalImpact} function from the package \href{https://google.github.io/CausalImpact/CausalImpact.html}{CausalImpact}.\cr
The sample with the largest (significant) \code{CausalImpact} is labeled
as the "event".
If \code{FALSE}, the sample where the change of sign (from negative to positive)
occurs is labeled as the "event".}

\item{Step}{An \code{integer} indicating the distance between samples to be assessed
when \code{findEvent = TRUE}.\cr
\code{Step} = 1 (default) means that all samples are considered during
the \code{findEvent} process.}

\item{chunk_size}{An \code{integer} defining the number of genes to be analysed
at a time. \code{chunk_size = 100} (default) indicates that groups of 100 genes
will be analised at a time.}

\item{PPItop}{A \code{numeric} value between 0 and 1 indicating the percentage of
PPI genes in the dataset to be selected as putative drivers. PPI genes with the
most interactions are selected. \code{PPItop = 0.3} (default) means that the 30% of the PPI
genes with the most interactions in the PPI network are selected.}

\item{alpha}{Significance level for the statistical test.
\code{alpha=0.05} by default.}

\item{CIniter}{number of iterations (200 by default)
for \code{CausalImpact} modeling.}

\item{returnModel}{If \code{TRUE}, the complete \code{CausalImpact} model is included
in the outcome of \code{findDCD}. If \code{FALSE} (default), only the most relevant
parameters of the \code{CausalImpact} model are included in the outcome of \code{findDCD}.}

\item{elbo_tol}{A \code{numeric} value (\code{elbo_tol = 1e-3} by default).
The relative pct change in the
evidence lower bound (ELBO) below which \code{phenopath} calculation of the
pseudotime score is considered converged.}

\item{project}{(optional) A TCGA project name (e.g. BRCA).
If provided, a dummy rank for the inferred dynamic cancer driver is calculated
based on the frequency of mutations of those genes in the TCGA project dataset.}
}
\value{
A \code{list} consisting of the following elements:
  \item{\code{res}}{A \code{list} with the results of the \emph{DynamicCancerDriver}
  inference process. Results are listed as follows:
  \enumerate{
          \item{\code{FS:}}{A \code{vector} containing the names of the putative
          cancer drivers}
          \item{\code{CausalImpact:}}{Causal impact models of the putative drivers}
          \item{\code{CDinfer:}}{Inferred Dynamic Cancer Drivers}
          \item{\code{summary:}}{A table with a summary of the results}
          }
          }
  \item{\code{eventAt}}{A \code{integer} containing the index (after pseudotime
  ordering) of the sample labeled as the "event".}
  \item{\code{z}}{Pseudotime score}
}
\description{
\code{findDCD} identifies genes driving driving one (or more)
significant biological processes along cancer progression based on the
hypothesis that the causal relationship between a cancer driver
gene and cancer development induces a significant deviation (also referred
as causal impact) of a core process from normal to carcinogenic.
}
\examples{
\dontrun{
   data("GSE75688_TPM_tumor", package = "DynamicCancerDriver")

----- Find Dynamic Cancer Drivers, PPI top 30\% -----
DCD.HER2time_SC <- findDCD(GeneExpression = GSE75688_TPM_tumor
                           , pathCovariate = "HER2"
                           , PPItop = 0.3
                           , findEvent = TRUE)
}

}
\references{

}
\seealso{
\link[DynamicCancerDriver]{findCovariate},
\link[DynamicCancerDriver]{parallelCI}
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
