% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findZ.r
\name{findZ}
\alias{findZ}
\title{findZ}
\usage{
function(GeneExpression, FS, pathCovariate, elbo_tol = 1e-3)\cr
}
\arguments{
\item{GeneExpression}{A \code{dataframe} (preferred) or a \code{matrix} containing mRNA gene expression.\cr
Columns represent mRNAs and rows represent samples. Column names should correspond
to gene IDs. Gene IDs can be from any of the following nomenclatures:
\emph{Ensembl.ID} (recommended), \emph{HGNC.ID, NCBI.ID} or \emph{HGNC.symbol}.}

\item{FS}{A \code{character} vector containing the names of the features
to be used for the calculation of the pseudotime score.}

\item{pathCovariate}{A \code{named vector} containing the data of a path
covariate.}

\item{elbo_tol}{A \code{numeric} value (\code{elbo_tol = 1e-3} by default).
The relative pct change in the
evidence lower bound (ELBO) below which \code{phenopath} calculation of the
pseudotime score is considered converged.}
}
\value{
A \code{dataframe} with the following two variables:
  \enumerate{
          \item{\code{Feature:}}{The vector \code{FS} of features}
          \item{\code{scontrol:}}{For each \code{Feature}, the name of the
          non-PPI gene with the largest Pearson correlation.}
          }
}
\description{
\code{findZ} calculates a pseudotime score for each sample
by using a \code{pathCovariate} that reasonably encodes the progression
of one core biological process during cancer progression (in the sense described by
\href{https://www.nature.com/articles/s41467-018-04696-6}{Campbell 2018}).\cr
Pseudotime score calculation relies in the procedures implemented in the
\href{https://www.bioconductor.org/packages/release/bioc/html/phenopath.html}{phenopath} package.
}
\examples{
\dontrun{
   data("GSE75688_TPM_tumor", package = "DynamicCancerDriver")
   FS <- colnames(GSE75688_TPM_tumor)[1:100]
   GE <- GSE75688_TPM_tumor[,1:500]
   sControl <- findCovariate(GeneExpression = GSE75688_TPM_tumor[,1:500]
   , FS = FS)

   #toy example, using "VIM" as path covariate
   z <- findZ(GeneExpression = GE
             , FS, pathCovariate = GSE75688_TPM_tumor[,"ENSG00000026025"]
             , elbo_tol = 1e-3)
             }


}
\references{

}
\seealso{
\link[DynamicCancerDriver]{findDCD}
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
