% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelCI.r
\name{parallelCI}
\alias{parallelCI}
\title{parallelCI}
\usage{
function(geneIDs = NULL, PPImatrix = NULL)\cr
}
\arguments{
\item{PPImatrix}{}
}
\value{
A \code{dataframe} consisting of the following elements:
  \item{\code{res}}{A \code{list} with the results of the DynamicCancerDriver
  inference process. Results are listed as follows:
  \enumerate{
          \item{\code{FS:}}{A \code{vector} containing the names of the putative
          cancer drivers}
          \item{\code{CausalImpact:}}{Causal impact models of the putative drivers}
          \item{\code{CDinfer:}}{Inferred Dynamic Cancer Drivers}
          \item{\code{summary:}}{A table with a summary of the results}
          }
   For each target gene with at least one parent. The index of the parents.}
  \item{\code{eventAt}}{A \code{integer} containing the index (after pseudotime
  order) of the sample labelled as the "event".}
  \item{\code{z}}{Pseudotime score}
}
\description{
parallelCI
}
\examples{
\dontrun{
   data("GSE75688_TPM_tumor", package = "DynamicCancerDriver")

----- Find Dynamic Cancer Drivers, PPI top 40\% -----
DCD.HER2time_SC <- findDCD(GeneExpression = GSE75688_TPM_tumor
                           , pathCovariate = "HER2"
                           , PPItop = 0.3
                           , findEvent = TRUE)
}

}
\references{

}
\seealso{
\link[DynamicCancerDriver]{findCovariate},
\link[DynamicCancerDriver]{parallelCI}
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
